==> Code /01_Data_Wrangling.Rmd <==
---
title: "Data Wrangling"
author: "Aye Nyein Thu"
date: "2024-12-08"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, warning=FALSE}
#Load packages 
library(tidyverse)
library(lubridate)
library(dplyr)
library(tidyr)
library(psych) #install.packages(psych) if not installed yet
library(knitr)
library(kableExtra)
library(plm)
library(car)
library(lmerTest)
library(leaflet)     
library(rnaturalearth)  
library(corrplot)
library(here)

#Check the working directory
here()
getwd()
```

```{r Data Importing and Initial Wrangling, include=FALSE, warning=FALSE}
#Raw Data 1: Fraser Institute's Economic Freedom Index 
#Import data
EconIns <- read.csv (file = here("./Data /Raw /FI_EconomicFreedom.csv"),
  stringsAsFactor = FALSE) #keep FALSE to avoid decimals counted as factors

#Clean data
EconIns_Processed <- EconIns %>%
  slice(-c(1:4)) %>% 
  select(X.1:X.4,X.19,X.30,X.39,X.57,X.80) %>%
  rename(Year = X.1,
         Country.Code = X.2,
         Country = X.3,
         Econ.Freedom = X.4,
         Govt.Size = X.19,
         Legal.Property = X.30,
         Sound.Money = X.39,
         Trade.Free = X.57,
         Regulation = X.80) %>% 
  mutate(across(-c(Country.Code, Country), as.numeric)) %>% #change from characters to numeric
  mutate(across(c(Country.Code, Country), as.factor)) #change from characters to factors

#Save to processed folder
write.csv(EconIns_Processed,
          file = ("./Data /Processed /EconIns_Processed.csv"),
          row.names = FALSE)

#Raw Data 2: CEPII's Geographic Data
#Import data
Geo <- read.csv (file = here("./Data /Raw /CEPII_GeographicClassification.csv"),
  stringsAsFactor = TRUE)

#Clean data 
Geo_Processed <- Geo %>% 
  select(iso3,country,continent,area,landlocked,lat,lon) %>% 
  distinct(iso3, .keep_all = TRUE) %>% 
  rename(Country.Code = iso3,
         Country = country,
         Continent = continent,
         Area = area,
         Landlocked = landlocked,
         Lat = lat, Lon = lon) %>% 
  mutate(across(-c(Country.Code, Country, Continent), as.numeric))

#Save to processed folder 
write.csv(Geo_Processed,
          file = ("./Data /Processed /Geo_Processed.csv"),
          row.names = FALSE)

#Raw Data 3: World Bank's World Governance Indicators 
#Import data
Gov <- read.csv(file = here("./Data /Raw /WB_WorldGovernance.csv"),
  stringsAsFactors = FALSE) #keep FALSE to avoid decimals counted as factors

#Clean data 
Gov_Processed <- Gov %>%
  slice(-c(5137:5178)) %>% 
  select(1:3,5,11,17,23,29,35) %>% 
  rename(Country = Country.Name,
         Year = Time,
         Corruption = Control.of.Corruption..Estimate..CC.EST.,
         Govt.Effectiveness = Government.Effectiveness..Estimate..GE.EST.,
         Political.Stability = Political.Stability.and.Absence.of.Violence.Terrorism..Estimate..PV.EST.,
         Voice.Accountability = Voice.and.Accountability..Estimate..VA.EST.,
         RuleofLaw = Rule.of.Law..Estimate..RL.EST.,
         Regulatory = Regulatory.Quality..Estimate..RQ.EST.) %>% 
  mutate(across(-c(Country.Code, Country), as.numeric)) %>% 
  mutate(across(c(Country.Code, Country), as.factor)) 

#Save to processed folder 
write.csv(Gov_Processed,
          file = ("./Data /Processed /Gov_Processed.csv"),
          row.names = FALSE)

#Raw Data 4: UNDP's Human Development Indicators (HDI)
#Import data
HDI <- read.csv (file = here("./Data /Raw /UNDP_HDI.csv"),
       stringsAsFactor = TRUE)

#Clean data 
HDI_Processed <- HDI %>%
  select(iso3, country, hdi_1990:hdi_2022) %>%
  rename_with(~ sub("hdi_", "", .), starts_with("hdi_")) %>%  #remove "hdi_" 
  pivot_longer(
    cols = `1990`:`2022`,  #change to panel data format using pivot longer
    names_to = "Year", values_to = "HDI") %>%
  rename(Country.Code = iso3, Country = country) %>% 
  mutate(Year = as.numeric(Year))  

#Save to processed folder 
write.csv(HDI_Processed,
          file = ("./Data /Processed /HDI_Processed.csv"),
          row.names = FALSE)

#Raw Data 5: World Bank's Income Classification
#Import data
WBClassify <- read.csv (
  file = here("./Data /Raw /WB_CountriesClassification_2024.csv"),
  stringsAsFactor = TRUE)

#Clean data 
WBClassify_Processed <- WBClassify %>% 
  select(Economy,Code,Region,Income.group) %>% 
  rename(Country = Economy,Country.Code = Code,Income.Group = Income.group) %>%
  mutate(Regional.Code = case_when(
      Region == "Latin America & Caribbean" ~ "LAC",
      Region == "South Asia" ~ "SA",
      Region == "Sub-Saharan Africa" ~ "SSA",
      Region == "Europe & Central Asia" ~ "ECA",
      Region == "Middle East & North Africa" ~ "MENA",
      Region == "East Asia & Pacific" ~ "EAP",
      Region == "North America" ~ "NAC")) %>% 
  slice(-c(219:265)) %>% 
  mutate(Regional.Code=as.factor(Regional.Code))

#Save to processed folder 
write.csv(WBClassify_Processed,
          file = ("./Data /Processed /WBClassify_Processed.csv"),
          row.names = FALSE)

#Raw Data 6: World Bank's Poverty and Inequality Platform (PIP)
#Import data
WBPIP <- read.csv (file=here("./Data /Raw /WB_PovertyInequalityPlatform.csv"),
       stringsAsFactor = TRUE)

#Clean data 
WBPIP_Processed <- WBPIP %>% 
  filter(reporting_level == "national") %>% 
  select(country_name,country_code,reporting_year,headcount,poverty_gap,
         poverty_severity) %>% 
  rename(Country = country_name,
         Country.Code = country_code,
         Year = reporting_year,
         Poverty.Headcount = headcount,
         Poverty.Gap = poverty_gap,
         Poverty.Severity = poverty_severity) %>% 
  mutate(Year = as.numeric(Year))

#Save to processed folder 
write.csv(WBPIP_Processed ,
          file = ("./Data /Processed /WBPIP_Processed.csv"),
          row.names = FALSE)

#Raw Data 7: World Development Indicators
#Import data
WDI <- read.csv(file=here("./Data /Raw /WDI_GrowthPopulation.csv"),
       stringsAsFactors = FALSE)

#Clean data 
WDI_Processed <- WDI %>% 
  select(-Time.Code) %>% 
  slice(-c(13889:17049)) %>% 
  mutate_all(~replace(., . == "..", NA)) %>% #replace ".,.." with NA
  rename(Country = Country.Name,
         Year = Time,
         RealGDP = GDP..constant.2015.US....NY.GDP.MKTP.KD.,
         GDP.Growth = GDP.growth..annual.....NY.GDP.MKTP.KD.ZG.,
         GDP.Capita = GDP.per.capita..constant.2015.US....NY.GDP.PCAP.KD.,
         Trade.Open = Trade....of.GDP...NE.TRD.GNFS.ZS.,
         Health.Exp = Current.health.expenditure....of.GDP...SH.XPD.CHEX.GD.ZS.,
         Education.Exp = Government.expenditure.on.education..total....of.GDP...SE.XPD.TOTL.GD.ZS.,
         Population.Growth = Population.growth..annual.....SP.POP.GROW.,
         Population.Density = Population.density..people.per.sq..km.of.land.area...EN.POP.DNST.,
         Population = Population..total..SP.POP.TOTL.) %>% 
  mutate(across(-c(Country.Code, Country), as.numeric)) %>% 
  mutate(across(c(Country.Code, Country), as.factor))

#Save to processed folder 
write.csv(WDI_Processed,
          file = ("./Data /Processed /WDI_Processed.csv"),
          row.names = FALSE)
```

```{r Family Data Wrangling, include=FALSE, warning=FALSE}
#Combine 7 processed datasets 
Combined <- WBPIP_Processed %>% 
  left_join(HDI_Processed, by=c("Country.Code","Year")) %>% 
  left_join(WDI_Processed, by=c("Country.Code","Year")) %>%
  left_join(Gov_Processed, by=c("Country.Code","Year")) %>%
  left_join(EconIns_Processed, by=c("Country.Code","Year")) %>%
  left_join(Geo_Processed, by=c("Country.Code")) %>%
  left_join(WBClassify_Processed, by=c("Country.Code")) %>%
  select (-Country.y,-Country.x.x,-Country.y.y,-Country.x.x.x,
          -Country.y.y.y,-Country) %>%
  rename (Country=Country.x) %>% 
  mutate(
    Governance = rowMeans(select(., Corruption:Voice.Accountability), 
                          na.rm = TRUE), #find the mean of governance indicators
  Gov.Expense = Health.Exp + Education.Exp) %>% #add governance expenditure
  select(Country:Year,Regional.Code,Region,Continent,
         Income.Group,Poverty.Headcount:Trade.Open,Gov.Expense,
         Health.Exp:Population,Governance,Corruption:Lon) 

#Filter the common years for all dataset 
Combined_CommonYear <- Combined %>% 
  filter(Year >= 1996 & Year <= 2021) 

#Drop all N/A values 
Combined_dropNA <- Combined_CommonYear %>% 
  drop_na()

#Save combined files in the processed folder
write.csv(Combined,
          file = ("./Data /Processed /Combined.csv"),
          row.names = FALSE)

write.csv(Combined_CommonYear,
          file = ("./Data /Processed /Combined_CommonYear.csv"),
          row.names = FALSE)

write.csv(Combined_dropNA,
          file = ("./Data /Processed /Combined_dropNA.csv"),
          row.names = FALSE)
```
