==> Code /02_Data Exploration.Rmd <==
---
title: "Data Exploration"
author: "Aye Nyein Thu"
date: "2024-12-09"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, warning=FALSE}
#Load packages 
library(tidyverse)
library(lubridate)
library(dplyr)
library(tidyr)
library(psych) #install.packages("psych") if not installed yet
library(knitr)
library(kableExtra)
library(plm)
library(car)
library(lmerTest)
library(leaflet)     
library(rnaturalearth)  
library(corrplot)
library(ggridges)
library(viridis)
library(RColorBrewer)
library(colormap)
library(ggthemes)
library(jtools) #install.packages("jtools") if not installed yet
library(ggcorrplot) #install.packages("ggcorrplot") if not installed yet
library(GGally)
library(stargazer)
library(here)

#Check the working directory
here()
getwd()

#Set ggplot theme
mytheme <- theme_economist() + 
  theme(
    plot.title = element_text(size = 12, face = "bold", hjust = 0.5),
    strip.text = element_text(size = 10, face = "bold"),  
    axis.title = element_text(size = 8),
    axis.text = element_text(size = 8),
    legend.position = "none" 
  )
theme_set(mytheme)
```

```{r Loading Data and Initial Data Exploration, warning=FALSE, message=FALSE}
#Import cleaned data without NA excluded
Ins_CommonYear <- read.csv(file="./Data /Processed /Combined_CommonYear.csv",
                     stringsAsFactors = TRUE) %>%
  mutate(across(c(Year, Population, Area, Landlocked), as.numeric))

#Import cleaned data with NA excluded
Ins_Panel <- read.csv(file="./Data /Processed /Combined_dropNA.csv",
                     stringsAsFactors = TRUE) %>%
  mutate(across(c(Year, Population, Area, Landlocked), as.numeric)) 

#Change to Cross-sectional data by taking yearly average
Ins_Cs <- Ins_Panel %>%
  group_by(Country, Country.Code) %>%
  summarise(
    across(where(is.numeric), mean, na.rm = TRUE), #return mean for numeric
    across(where(is.factor), ~ first(.)), #return first value for factor
  ) %>% 
  select(-Year) %>% #remove year as it is no longer relevant
  select(Country,Country.Code,
         Regional.Code:Income.Group,everything()) #rearrange columns 

#Save the Cross-sectional data in the processed folder 
write.csv(Ins_Cs,
          file = ("./Data /Processed /Ins_Cs.csv"),
          row.names = FALSE)
```

```{r Data Structure, message=FALSE, warning=FALSE}
#Check the data structure 
dim(Ins_Cs)
summary(Ins_Cs)
```

```{r Summary Variables}
VNames <- c("Poverty Headcount", "Human Development Index", 
          "Gross Domestic Product (GDP) per capita (Constant 2015 US$)", 
          "Governance", "Economic Freedom", "Government Expenditure",
          "Population", "Landlocked","Area")

CNames <- c("Poverty.Headcount", "HDI", "GDP.Capita", "Governance", 
            "Econ.Freedom", "Gov.Expense", "Population", "Landlocked","Area")

Units <- c("% of Population", "Score ranging from 0 (lowest) to 1 (highest)",
           "USD", "Score ranging from -2.5 (lowest) to 2.5 (highest)", 
           "Score ranging from 0 (lowest) to 10 (highest)", "% of GDP", 
           "No. of People", "0(not landlocked) or 1 (landlocked)", 
           "Square kilometers") 

Sources <- c("World Bank's Poverty and Inequality Platform (PIP)",
             "United Nations Development Programme (UNDP)",
             "World Bank's World Development Indicators (WDI)",
             "World Bank's Worldwide Governance Indicators",
             "Fraser Institute (FI)",
             "World Bank's World Development Indicators (WDI)",
             "World Bank's World Development Indicators (WDI)",
             "CEPII",
             "CEPII")

Definition <- c("% of people residing in households with consumption or income per individual falling below the international poverty threshold in 2017",
                "HDI is a development index that combines 3 key indicators: [1] Health by life expectany, [2]Education by total schooling years and [3] Standards of living by GNI per capita.",
                "GDP is the total value additions of producers residing in the nations. It adds product taxes and deducts subsidies.",
                "Governance indicators are composed of 6 dimensions: [1] Voice and Accountability; [2] Political Stability and Absence of Violence/ Terrorism; [3] Government Effectiveness; [4] Regulatory Quality; [5] Rule of Law; and [6] Control of Corruption. The study aggregates these 6 dimensions as Governance Indicators by finding average.",
                "The Economic Freedom Index by Fraser Institute measures how much economic freedom a country has by using 5 different areas: [1] Size of Government; [2] Legal System and Security of Property Rights; [3] Sound Money; [4] Freedom to Trade Internationally; [5] Regulation.",
                "The study combines total government expenditure on education and current health expenditure to define as Government Expenditure.",
                "Total population based on estimates of national population censuses, fertility, mortality and migration",
                "It represents the geographic condition if a country is landlocked or not.",
                "It represents the geographic area of a country in square kilometers.")

datast <- data.frame(VNames, CNames, Units, Sources, Definition)
colnames(datast) <- c("Variable Names", "Column Names", "Units","Sources",
                      "Definition")
  
knitr::kable(datast, caption ="Summary of Variables under Study") 

```

```{r Summary Statistics}
options(scipen = 10)
#Find summary statistics of variables used 
Summary <- Ins_Panel %>% 
  select(Poverty.Headcount,GDP.Capita,HDI,
         Gov.Expense,Population,Area,
         Governance,Econ.Freedom) %>% 
  summarise(across(everything(), list(
    count = ~sum(!is.na(.)),
    mean = ~mean(., na.rm = TRUE),
    sd = ~sd(., na.rm = TRUE),
    min = ~min(., na.rm = TRUE),
    max = ~max(., na.rm = TRUE),
    range = ~diff(range(., na.rm = TRUE)),
    se = ~sd(., na.rm = TRUE) / sqrt(sum(!is.na(.)))
  )))

#Format the summary statistics 
Summary_Formatted <- Summary %>%
  pivot_longer(everything(), 
               names_to = c("variable", "statistic"), 
               names_sep = "_") %>%
  pivot_wider(names_from = statistic, values_from = value) 

Summary_Formatted %>%
  kable("html", booktabs = TRUE, caption = "Descriptive Statistics",
        digits = 2) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"), 
                full_width = FALSE)
```

```{r Overview of Institutions, warning = FALSE, message=FALSE}
#Overview of Governance in the countries under study
#Retrieve world geographic data and change to sf format 
World <- ne_countries(returnclass = "sf") 

#Join the cleaned data with geographic data by country code 
Map_Data <- World %>%
  left_join(Ins_Cs, by = c("iso_a3" = "Country.Code"))

#Set the color palette 
Pal.Gov <- colorNumeric(palette = "RdYlBu", domain = Map_Data$Governance)

#Plot Governance of the countries under study
leaflet(Map_Data) %>%
  addTiles() %>% 
  addPolygons(
    fillColor = ~Pal.Gov(Governance),  
    fillOpacity = 0.7,
    color = "white",
    weight = 1,
    popup = ~paste(name, "<br>Governance: ", Governance) 
  ) %>%
  addLegend(
    position = "bottomright",
    pal = Pal.Gov, 
    values = ~Governance,
    title = "Governance",
    opacity = 0.7
  )

#Overview of Economic Freedom in the Countries under Study
#Set the color palette 
Pal.Econ <- colorNumeric(palette = "YlGnBu", domain = Map_Data$Econ.Freedom)

#Plot Economic Freedom of the countries under study
leaflet(Map_Data) %>%
  addTiles() %>% 
  addPolygons(
    fillColor = ~Pal.Econ(Econ.Freedom),  
    fillOpacity = 0.7,
    color = "white",
    weight = 1,
    popup = ~paste(name, "<br>Econ.Freedom: ", Econ.Freedom) 
  ) %>%
  addLegend(
    position = "bottomright",
    pal = Pal.Econ,  
    values = ~Econ.Freedom,
    title = "Economic Freedom",
    opacity = 0.7
  )
```

```{r Relationship of Main Variables, warning = FALSE, message=FALSE}
#Plot the relationship between Poverty and Governance across regions
ggplot(Ins_Cs, aes(x = Governance, y = Poverty.Headcount, fill = Continent)) +
  geom_point(shape = 21, size = 3, color = "black", stroke = 0.5, alpha = 0.7) +  
  geom_smooth(method = "lm", se = FALSE, color = "red", linetype = "dashed") + 
  scale_fill_brewer(palette = "Set2") +  
  labs(
    title = "Poverty vs. Governance by Regions",
    x = "Governance (scale from -2.5[lowest] to 2.5[highest])",
    y = "Poverty Headcount (% of population)"
  ) +
  facet_wrap(~Continent, scales = "free")

#Plot the relationship between Growth and Governance across regions
ggplot(Ins_Cs, aes(x = Governance, y = GDP.Capita, fill = Continent)) +
  geom_point(shape = 21, size = 3, color = "black", stroke = 0.5, alpha = 0.7) +  
  geom_smooth(method = "lm", se = FALSE, color = "red", linetype = "dashed") + 
  scale_fill_brewer(palette = "Set2") +  
  labs(
    title = "Growth vs. Governance by Regions",
    x = "Governance (scale from -2.5[lowest] to 2.5[highest])",
    y = "GDP per capita (2015 constant USD)"
  ) +
  facet_wrap(~Continent, scales = "free")

#Plot the relationship between HDI and Governance across regions
ggplot(Ins_Cs, aes(x = Governance, y = HDI, fill = Continent)) +
  geom_point(shape = 21, size = 3, color = "black", stroke = 0.5, alpha = 0.7) +  
  geom_smooth(method = "lm", se = FALSE, color = "red", linetype = "dashed") + 
  scale_fill_brewer(palette = "Set2") +  
  labs(
    title = "HDI vs. Governance by Regions",
    x = "Governance (scale from -2.5[lowest] to 2.5[highest])",
    y = "Human Development Index (HDI) (scale from 0[lowest] to 1[highest])"
  ) +
  facet_wrap(~Continent, scales = "free") 

#Plot the relationship between Main Development Variables and Economic Freedom
#Add 3 main dependent variables into pivot_longer frame
Dependent_Variables <- Ins_Cs %>%
  pivot_longer(
    cols = c(GDP.Capita, Poverty.Headcount, HDI), 
    names_to = "Variable",                        
    values_to = "Value"                          
  )

#Plot
ggplot(Dependent_Variables, aes(x = Econ.Freedom, y = Value)) +
  geom_point(alpha = 0.7) +  
  geom_smooth(method = "lm", se = FALSE, color = "blue") + 
  facet_wrap(~Variable, scales = "free_y", nrow = 1) +  
  labs(
    title = "Main Development Variables vs. Economic Freedom",
    x = "Economic Freedom (scale from 0[lowest] to 10[highest])",
    y = "Value"
  )
```

```{r Correlation Plot}
#Correlation Plot
Data_Cor <- Ins_Cs %>% 
  select(Poverty.Headcount, GDP.Capita, HDI,
         Gov.Expense, Population, Landlocked, Area, 
         Governance, Econ.Freedom) 
Numeric_Data <- Data_Cor[, sapply(Data_Cor, is.numeric)] #ensure all are numeric

Correlation <-  cor(Numeric_Data)

ggpairs(Numeric_Data) + 
  labs(title = "Correlation Analysis") + 
  theme_gray() + 
    theme(
    plot.title = element_text(size = 14, face = "bold", hjust = 0.5),
    strip.text = element_text(size = 8),   
    axis.text = element_text(size = 8),  
    axis.title = element_text(size = 8),  
    legend.text = element_text(size = 8),  
    legend.title = element_text(size = 8)  
  )
```
